# docs/swagger.yaml
basePath: /
definitions:
...
  catalog.CatalogItemResponse:
    properties:
      attributes:
        additionalProperties: true
        type: object
      base_price:
        example: 25.5
        type: number
      category:
        example: vegetables
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      currency:
        example: INR
        type: string
      description:
        example: Fresh organic tomatoes grown without pesticides
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      images:
        example:
          - https://example.com/tomato1.jpg
          - https://example.com/tomato2.jpg
        items:
          type: string
        type: array
      is_active:
        example: true
        type: boolean
      item_type:
        example: PRODUCT
        type: string
      name:
        example: Organic Tomatoes
        type: string
      organization_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      sku:
        example: TOM-ORG-001
        type: string
      subcategory:
        example: tomatoes
        type: string
      tags:
        example:
          - organic
          - fresh
          - local
        items:
          type: string
        type: array
      unit_of_measure:
        example: kg
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      visibility:
        example: ORG
        type: string
    type: object
...
  catalog.PriceUpdateRequest:
    properties:
      base_price:
        minimum: 0
        type: number
      currency:
        type: string
      effective_date:
        type: string
      promotion_end_date:
        type: string
      promotional:
        type: boolean
      reason:
        maxLength: 500
        type: string
    required:
      - base_price
    type: object
...
  catalog.PublishRequest:
    properties:
      effective_date:
        type: string
      notes:
        maxLength: 500
        type: string
      visibility:
        enum:
          - PRIVATE
          - ORG
          - NETWORK
          - PUBLIC
        type: string
    type: object
...
  catalog.UnpublishRequest:
    properties:
      effective_date:
        type: string
      handle_active_orders:
        enum:
          - fulfill
          - cancel
          - transfer
        type: string
      reason:
        maxLength: 500
        type: string
    type: object
...
  common.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      error:
        $ref: "#/definitions/common.ResponseError"
      meta:
        $ref: "#/definitions/common.ResponseMeta"
    type: object
  common.ResponseError:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  common.ResponseMeta:
    properties:
      extra:
        additionalProperties: true
        type: object
      pagination:
        $ref: "#/definitions/common.PaginationMeta"
      timestamp:
        type: string
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the KisanLink E-commerce service API.
  title: KisanLink E-commerce API
  version: "1.0"
paths:
...
  /api/v1/catalog/{type}/{id}/price:
    put:
      consumes:
        - application/json
      description: Update the price of a catalog item with price history tracking
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Item type (products, services, labour, contracts)
          in: path
          name: type
          required: true
          type: string
        - description: Item ID
          in: path
          name: id
          required: true
          type: string
        - description: Price update data
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/catalog.PriceUpdateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/catalog.CatalogItemResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Update catalog item price
      tags:
        - Catalog Management
  /api/v1/catalog/{type}/{id}/publish:
    post:
      consumes:
        - application/json
      description:
        Publish a catalog item to make it visible according to its visibility
        settings
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Item type (products, services, labour, contracts)
          in: path
          name: type
          required: true
          type: string
        - description: Item ID
          in: path
          name: id
          required: true
          type: string
        - description: Publish options
          in: body
          name: request
          schema:
            $ref: "#/definitions/catalog.PublishRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/catalog.CatalogItemResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "409":
          description: Conflict
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Publish catalog item
      tags:
        - Catalog Management
  /api/v1/catalog/{type}/{id}/unpublish:
    post:
      consumes:
        - application/json
      description: Unpublish a catalog item to make it inactive
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Item type (products, services, labour, contracts)
          in: path
          name: type
          required: true
          type: string
        - description: Item ID
          in: path
          name: id
          required: true
          type: string
        - description: Unpublish options
          in: body
          name: request
          schema:
            $ref: "#/definitions/catalog.UnpublishRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/catalog.CatalogItemResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "409":
          description: Conflict
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Unpublish catalog item
      tags:
        - Catalog Management
...
swagger: "2.0"